Implemented a user required conflict resolution when files where edited locally AND in the server.  The files will not be updated until the user selects which of the two files is the newest one. Note that at this time, other clients may be updating the server. So the client should take caution to select the proper new file.
The server file is downloaded to disk and prepended with _ocs_conflict.
so that the user can open both files and compare accordingly.

A message appears notifying the user of the conflict, and the clouds
turn red. They will not be reverted back to blue until all conflicts are
cleared.