Add doc cmake targets (HTML, PDF, QtHelp, CHM, man)
make doc will build all of the above, except for CHM,
which needs manual preparation and can be built with
make doc-chm. See doc/scripts/README.rst for details.

We do our best to ensure to detect the required tools
before adding targets, so a build should always succeed.

Exception: On Debian and Ubuntu, the following packages are
required to build the PDF target (in addition to pdflatex
itself, which is autodetected):

* texlive-latex-recommended
* texlive-latex-extra
* texlive-fonts-recommended

If pdflatex is present, but those are not, the doc target
will fail.

Results can be found in $BUILDDIR/doc/$format.