Consolidate theming support.
This implements out-of-repository-theming, and removes the old
custom.ini theming that never worked on all platforms and had
no straight deployment story.

The new approach requires the CMAKE variable OEM_THEME_DIR to
point to a directory that must at least contain an OEM.cmake file
(check OWNCLOUD.cmake for possible options) as well as a themes.qrc
and a themes/ directory that directly correspond to the ones in
the source tree.