Don't cleanup in owncloud's vio_module_shutdown
All the cleanup is done in owncloud_commit which should
necessarily be called.

Should fix https://github.com/owncloud/mirall/issues/713
The problem was that csync_destroy was called on another csync
instance in another thread, resulting in a crash

Yes, there is still a problem with the proxy settings.