Don't read from the db if the file id has changed.
File id may change when upgrading to owncloud 6 or when upgrading to mirall 1.5

This also changes the location where we check if we need to read from
the DB.   Instead of doing it from csync_ftw,  we do it from
_csync_detect_update. In that function we know the etag and the file id.
Which mean we change the flag before the recursive call to csync_ftw instead
of in the begin of the function.  Therefore, we need to restore the flag
after calling the function to the previously saved state.