Differentiate the RENAME and EVAL_RENAME
EVAL_RENAME is set at the update phase and means that this is the destination of
a file that has been renamed.
RENAME is set for the file that need to be renamed.

Before, we knew the difference because of which remote the instruction was
but now we need to make the distinction explicit