Do not set sync-state to Prepare in setSyncEnabled().
This method is also called if the sync was disabled due to
network problems and such, ie. from the ETag change check.
In that case, we do not want to go to the Prepare state as
we not neccessarily want to have the folder synced now.

This fixes the problem that the folders were resynced way
to often which we had in master for a couple of days now.