Sharedialog: Minor fixes
The server generates a new share when a password is set/deleted so we need to
refetch the shares once that is done.

It is not possible to set an expiration date in the past for a share. So do not
allow the user to select a date in the past.