Folder: Generate an alias for new folders
Before commit 1a51b6718a345f1e5f28dd1f4e0c04bdaeeb5357, the wizard was
making sure folder had an alias but this is no longer the case.
So generate still an unique alias.

Alias is not used in the UI any longer, it's just use for internal purposes.

For issue #4737