Revert "Fix a deadlock when shutting down during discovery" (#5100)
Reverts commit 622017adcf3b1a914221acc0bb2c17e1429ea4c3

Could be the cause of #5092 and the cost is higher than the benefit if this is the case.
A network request taking more than 30 seconds isn't something unlikely in this world
and shouldn't be a good reason to abort. We should try to untangle the threads
dependencies to properly fix this if possible instead.