delta-sync Fix new resuming tests after merge
Adjust TestChunkingNG::testResume3 and TestChunkingNG::testResume4 to the fact
that the name of the chunks is now the offset.

Note: testResume4 used to discard the whole directory with such inconsistency.
We do no longer do that. That's because PropagateUploadFileNG::slotPropfindFinished
check each chunk with 'markRangeAsDone', and just discar individual chunks
that are outside of the range. And a chunk that would overlap would necessarily
be outside of the ranges.