Sharing: Create shares with appropriate permissions #7275
If user A shares something with user B and allows resharing but removes
some other permission then user B should be able to share the item or
any of its contents with user C and the new share permissions should
automatically be <= the permissions of the original share.

Currently this works correctly when resharing the original item but
fails for any subitem, preventing the reshare entirely.

With this patch the reshare will always be created with appropriately
limited permissions.