Fix leak in CloudProviderManager
The member _map, allocated in the constructor, was leaking.

The ad-hoc way to fix the leak would have been to add a destructor
that deletes _map.  However, there is no good reason why _map has
to be a pointer to a QMap (instead of a QMap).
Hence this commit changes the type of the _map member from a QMap
pointer to a QMap.

Signed-off-by: Stephan Beyer <s-beyer@gmx.net>
(cherry picked from commit eaf4aa0e5b5d17b2061aa5ae9d0edce968f5a1e5)
Signed-off-by: Michael Schuster <michael@schuster.ms>