Replace last foreach invocation with while loop.
Fix the last remaining foreach invocation in OCC::SqlDataBase::close.

Since the container is being modified while iterating over it's members
using a range based for or other iterator based looping construct is not
safe. I've replaced it by a while loop, calling the finish method on
what at any time is the first item in the container.

Signed-off-by: Harald Eilertsen <haraldei@anduin.net>