Display native language names instead of locale names
This is a lot more user friendly than having to work with locale names, which often can't easily be associated with the language they represent.

There is also a fallback for cases when Qt cannot provide the native name for a locale provided by the translations helper. This usually indicates an issue in the filename of the translations. Right now, this is the case with "client_TW.qm".