Tweak the settings dialog size to correctly show the wizards
With the style we use, the wizard is sized to "fit" inside the settings
dialog, and it will also be restrainded to the size of the settings
dialog.

This change makes sure that on macOS, the settings dialog has enough
(vertical) space to show all the widgets in all pages. If there is not
enough space, widgets might overlapped by other widgets (or disappear
behind them).

Fixes: #8919