CMake: use ccache and sccache for compiler caching
First build on Fedora 35 with an Intel Core i7 8550U: 4.5 minutes

Repeat build with ccache after deleting CMake build directory
and creating a new one: 1.5 minutes

This code is adapted from Tenacity, written by myself and
@emabrey, licensed BSD-3-Clause.

ccache is easily available from Linux distribution packages and
Homebrew on macOS. sccache is the only currently maintained
compiler cache for MSVC and can be installed easily from
Chocolatey on Windows. Alternatively it can be installed with
cargo if a Rust toolchain is installed.

Signed-off-by: Be <be@mixxx.org>