Split up controller logic into multiple classes
This is a mix of the state and abstract factory patterns. Moving the functionality into multiple classes reduces the complexity of the controller code and makes adding new states easier.

The controller is still in charge of deciding when to display which page etc., and creates the single objects. It is, however, no longer in charge of creating the page widgets, this is handled by the state objects. While this removes the dependency injection functionality used elsewhere within the new wizard, currently, there is no advantage in having the controller be fully in charge of the creation of the widgets.