cloudproviders keep track of the length of the main menu items
cloudproviderwrapper creates (and provides a menu model) that is used
to populate a menu within file managers.  The menu includes a
submenu of recent activity that has no limit on the maximum number of
characters.  In order to create a more reasonably sized menu,
let's keep track of the width of the items in the main menu as a
baseline for the width of the items in the submenu.  This is
done dynamically in order to account for translations

Signed-off-by: Corey Berla <corey@berla.me>