Debian build classification 'beta' cannot override 'release'.
The Ubuntu/Debian build script classifies the built commit based
on the tags, if any, pointing to the commit. This classification
is used to determine the target PPA.

If no suitable tag points the commit, it will be an 'alpha'
build. If a tag of the format 'vA.B.C' points to the commit,
it will be a 'release' build, if the version is followed by
something, e.g. 'vA.B.C-rc1', it will be a 'beta' build.

The current classification fails if both a 'beta' and
a 'release' tag points to the commit, such as in case of
the recent v3.7.4 release, where the '-rc1' tag points to the
commit that was later released. It was deemed to be a 'beta'
build. This patch fixes this behaviour to build such commits
as 'release'es.

Signed-off-by: István Váradi <ivaradi@varadiistvan.hu>