Fix crash when database version check fails
The check is done after the application constructor is run. In this
constructor, a couple of signals get fired, which in turn queues showing
the GUI window (e.g. an account is logged out, and needs new
authentication). When the version check fails, it brings up a (modal)
dialog, and the event loop is spun, which also brings up the GUI window.
Somewhere halfway this, the quit button is clicked, tearing down the
application with a half-way set up window. This causes a crash somewhere
deep in a child-widget-deletion process.

The fix is to move even more out of the application constructor, so the
full check can be done before any other GUI code is run. This comes down
to setting up the account manager and the folder manager in the main
function.

Fixes: #11135