fixed detection of server version in cmd client similar to gui client:
first check status.php for version and then capabilities, only use
version string if not empty

this also fixes 'File names containing the character ":" are not
supported on this file system.' errors in nextcloudcmd (on Linux):
The invalidFilenameRegex was set to a static default in case the server
version was not set correctly. As newer versions of nextcloud do not
return the version in capabilities but status.php, the server version
was empty.