Change when checking for updates
We now handle user-requested actions from the command-line, and not
check for updates, because a user would not appreciate it if the
application quits because an update is found.

Also move opening the new-account wizard to after the update check, so
that when no account is configured, and an update is available, only
one dialog will be shown.

Fixes: #11141