fix window resize on DPI changes
This commit fixes #7678 and likely fixes #7447 too.  On Windows, when
the screen is changed to one with a different scaling setting the window
is being incorrectly resized.  This can be worked around by storing the
window's size before moving it to a different screen, and then resizing
the window again.

Signed-off-by: Jyrki Gadinger <nilsding@nilsding.org>