persist credentials retrieved through the wizard
When setting up an account through the wizard for the first time, the
credentials are not saved properly.  This requires the user to sign in
again after an application restart.

At least for the login flow v2 case, the `WebFlowCredentials` class
eventually persists the retrieved credentials, whereas the class used
within the wizard (`OCC::Flow2AuthCredsPage`) doesn't.

Signed-off-by: Jyrki Gadinger <nilsding@nilsding.org>