feat: build translated MSI installer
the base translations come from WixUI itself

There are still some untranslated strings in Nextcloud.wxs, we need to
take care of them at a later point.  As a first step this is good
enough.

This commit serves as a proof-of-concept, the languages I used here will
change depending on what we want to support.

Signed-off-by: Jyrki Gadinger <nilsding@nilsding.org>